Refactoring file watcher.
