+++ src/tools/os/linux/file/watcher.cc
@@ -17,196 +17,136 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
- #include "private.h"
- #include <unistd.h>
- #include <cstring>
- #include <udjat/tools/threadpool.h>
- #include <udjat/tools/file.h>
- #include <sched.h>
+ /**
+  * @brief Implements linux file watcher.
+  */
+
+ #include <config.h>
+ #include <private/filewatcher.h>
+ #include <sys/inotify.h>
+ #include <udjat/tools/logger.h>
  #include <sys/types.h>
  #include <sys/stat.h>
- #include <unistd.h>
- #include <sys/types.h>
- #include <sys/stat.h>
- #include <fcntl.h>
- #include <iostream>
+ #include <stdexcept>
+
+ #ifdef HAVE_UNISTD_H
+	#include <unistd.h>
+ #endif // HAVE_UNISTD_H
+
+ #define INOTIFY_EVENT_SIZE ( sizeof (inotify_event) )
+ #define INOTIFY_EVENT_BUF_LEN ( 1024 * ( INOTIFY_EVENT_SIZE + 16 ) )
 
  using namespace std;
 
  namespace Udjat {
 
-	std::mutex File::Watcher::guard;
 
-	File::Watcher::Watcher(const Quark &n) : name(n) {
-		Controller::getInstance().insert(this);
-	}
+	File::Watcher::Controller::Controller() {
 
-	File::Watcher::~Watcher() {
-		Controller::getInstance().remove(this);
-	}
+		Logger::String{"Starting file watcher"}.trace("services");
 
-	File::Watcher * File::watch(void *id, const char *name, std::function<void (const Udjat::File::Text &)> callback) {
+		MainLoop::Handler::fd = inotify_init1(IN_NONBLOCK|IN_CLOEXEC);
+		if(MainLoop::Handler::fd == -1) {
+			throw system_error(errno,system_category(),"Can't initialize inotify");
+		}
 
-		Watcher * watcher =  Controller::getInstance().find(name);
-		watcher->push_back(id,callback);
-		return watcher;
+		MainLoop::Handler::events = MainLoop::Handler::oninput;
+		MainLoop::Handler::enable();
 
 	}
 
-	File::Watcher * File::watch(void *id, const Quark &name, std::function<void (const Udjat::File::Text &)> callback) {
+	File::Watcher::Controller::~Controller() {
 
-		Watcher * watcher =  Controller::getInstance().find(name.c_str());
-		watcher->push_back(id,callback);
-		return watcher;
-
-	}
-
-	void File::Watcher::remove(void *id) {
 		std::lock_guard<std::mutex> lock(guard);
 
-		files.remove_if([id](const File &file) {
-			return file.id == id;
-		});
+		/*
+		for(auto watcher : watchers) {
 
-		if(files.size())
-			return;
-
-		delete this;
-	}
-
-	bool File::Watcher::update(bool force) {
-
-		if(updated && !force) {
-			return false;
-		}
-
-		try {
-
-#ifdef DEBUG
-			cout << "Inotify\tLoading '" << name.c_str() << "' for " << files.size() << " file(s)" << endl;
-#endif // DEBUG
-
-			// Open file
-			int fd = open(name.c_str(),O_RDONLY);
-			if(fd < 0) {
-				throw system_error(errno, system_category(), (string{"Can't open '"} + name.c_str() + "'"));
+			if(watcher->wd != -1) {
+				inotify_rm_watch(MainLoop::Handler::fd, watcher->wd);
+				watcher->wd = -1;
 			}
 
-			try {
+		}
+		*/
 
-				struct stat st;
-				if(fstat(fd, &st)) {
-					throw system_error(errno, system_category(), (string{"Can't get stats of '"} + name.c_str() + "'"));
-				}
+		MainLoop::Handler::disable();
+		::close(MainLoop::Handler::fd);
 
-				if(st.st_mtime == this->mtime) {
-#ifdef DEBUG
-					cout << "File was not modified" << endl;
-#endif // DEBUG
-					::close(fd);
-					return false;
-				}
+	}
 
-				this->mtime = st.st_mtime;
+	File::Watcher::Controller & File::Watcher::Controller::getInstance() {
+		static File::Watcher::Controller instance;
+		return instance;
+	}
 
-				if(files.size()) {
+	void File::Watcher::Controller::insert(File::Watcher *watcher) {
 
-					Udjat::File::Text file(fd,st.st_size);
+		struct stat st;
+		if(stat(watcher->pathname, &st)) {
+			throw std::system_error(errno,std::system_category(),watcher->pathname);
 
-					for(auto f = files.begin(); f != files.end(); f++) {
+		}
 
-						try {
+		if((st.st_mode & S_IFDIR)) {
 
-							f->callback(file);
+			// Directory watch
+			watch_directory(watcher->pathname,watcher);
 
-						} catch(const exception &e) {
+		} else if( (st.st_mode & S_IFREG)) {
 
-							cerr << "inotify\tError '" << e.what() << "' updating file '" << this->name.c_str() << "'" << endl;
+			// File watch.
+			watch_file(watcher->pathname,watcher);
 
-						}
+		} else if( (st.st_mode & S_IFLNK)) {
 
-					}
+			// Symbolic link
+			throw runtime_error(Logger::String{"Unable to watch symlink '",watcher->pathname,"'"});
 
-				}
+		} else {
 
-			} catch(...) {
+			throw runtime_error(Logger::String{"File '",watcher->pathname,"' has unexpected type"});
 
-				::close(fd);
-				throw;
+		}
+
+	}
+
+	void File::Watcher::Controller::watch_file(const char *path, File::Watcher *watcher) {
+	}
+
+	void File::Watcher::Controller::watch_directory(const char *path, File::Watcher *watcher) {
+	}
+
+	void File::Watcher::Controller::remove(File::Watcher *watcher) {
+	}
+
+	void File::Watcher::Controller::onEvent(const ::inotify_event *event) noexcept {
+
+	}
+
+	void File::Watcher::Controller::handle_event(const MainLoop::Handler::Event) {
+
+		char * buffer = new char[INOTIFY_EVENT_BUF_LEN];
+		memset(buffer,0,INOTIFY_EVENT_BUF_LEN);
+
+		ssize_t bytes = read(buffer, INOTIFY_EVENT_BUF_LEN);
+
+		while(bytes > 0) {
+
+			ssize_t	bufPtr	= 0;
+
+			while(bufPtr < bytes) {
+				const ::inotify_event *pevent = (::inotify_event *) &buffer[bufPtr];
+				onEvent(pevent);
+				bufPtr += (offsetof (::inotify_event, name) + pevent->len);
 			}
 
-			::close(fd);
-
-		} catch(const exception &e) {
-
-			cerr << "inotify\tError '" << e.what() << "' loading '" << this->name.c_str() << "'" << endl;
-
+			bytes = read(buffer, INOTIFY_EVENT_BUF_LEN);
 		}
 
-		return updated = true;
+		delete[] buffer;
 
 	}
-
-	void File::Watcher::onChanged() noexcept {
-
-		Controller::getInstance().remove(this);
-
-		ThreadPool::getInstance().push("FileWatcher",[this]() {
-
-			sched_yield();
-
-			if(::access(this->name.c_str(),F_OK) == -1) {
-
-				cerr << "inotify\tFile '" << name.c_str() << "': " << strerror(errno) << endl;
-
-			} else {
-
-				update(true);
-				Controller::getInstance().insert(this);
-
-			}
-
-		});
-
-	}
-
-	void File::Watcher::onEvent(const uint32_t event) noexcept {
-
-		if(event & IN_IGNORED) {
-
-			// Watch  was  removed  explicitly or automatically (file  was  deleted, or filesystem was unmounted)
-			cout << "inotify\tFile '" << this->name.c_str() << "' was ignored" << endl;
-			this->wd = -1;
-			onChanged();
-		}
-
-		if(event & IN_CLOSE_WRITE) {
-			cout << "inotify\tFile '" << this->name.c_str() << "' was changed" << endl;
-			ThreadPool::getInstance().push("FileWatcherEvent",[this]() {
-				sched_yield();
-				update(true);
-				Controller::getInstance().insert(this);
-			});
-		}
-
-		if(event & IN_DELETE_SELF) {
-			cout << "inotify\tFile '" << this->name.c_str() << "' was deleted" << endl;
-		}
-
-		if(event & IN_MOVE_SELF) {
-			cout << "inotify\tFile '" << this->name.c_str() << "' was moved" << endl;
-			onChanged();
-		}
-
-	}
-
-	void File::Watcher::push_back(void *id, std::function<void (const Udjat::File::Text &)> callback) {
-
-		std::lock_guard<std::mutex> lock(guard);
-		File file(id,callback);
-		files.push_back(file);
-
-	}
-
 
  }
+
